split state to component files
