+++ crates/revm/src/db/states/account_status.rs
@@ -1,0 +1,15 @@
+
+#[derive(Clone, Default, Debug)]
+pub enum AccountStatus {
+    #[default]
+    LoadedNotExisting,
+    Loaded,
+    LoadedEmptyEIP161,
+    Changed,
+    New,
+    NewChanged,
+    Destroyed,
+    DestroyedNew,
+    DestroyedNewChanged,
+    DestroyedAgain,
+}
